rag_service/rag_system/requirements.txt
# RAG System Dependencies
openai>=1.3.0
faiss-cpu>=1.7.4
numpy>=1.24.0
asyncio-compat>=0.1.2
 
# Optional dependencies
# sentence-transformers>=2.2.2  # For local embeddings
# ollama>=0.1.0  # For local fallback 